// src/app/interfaces/journal.ts
export interface Journal {
    journalId?: string
    userId?: string
    todaysFocus:  string
    excitedAbout?: string
    todaysGoal?: string
    affirmation?: string
    goodThings1?: string
    goodThings2?: string
    goodThings3?: string
    positiveActions1?: string
    positiveActions2?: string
    positiveActions3?: string
    gratefulFor1?: string
    gratefulFor2?: string
    gratefulFor3?: string
    peopleMadeFeelGood1?: string
    peopleMadeFeelGood2?: string
    peopleMadeFeelGood3?: string
    createdAt?: Date
    updatedAt?: Date
}